(The receiver should not process voice command through orbsessioncallback) removed in HTM-740 Send data directly to orbsession Renaming the voice command function Update onVoiceSendKeyAction to dispatch key event Remove dispatchKeyPressEvent and notifyKeyUp Move voice service receiver to TIS (as well as mainactivity in mockapp) Create interface default method for consoleLog in session callback
