+++ android/orblibrary/src/main/java/org/orbtv/orblibrary/OrbSession.java
@@ -1,6 +1,7 @@
 package org.orbtv.orblibrary;
 
 import android.content.Context;
+import android.content.IntentFilter;
 import android.os.Handler;
 import android.os.Looper;
 import android.util.Log;
@@ -15,7 +16,6 @@ import java.util.List;
 
 class OrbSession implements IOrbSession {
     private static final String TAG = OrbSession.class.getSimpleName();
-
     private final IOrbSessionCallback mOrbSessionCallback;
     private final int mOrbHbbTVVersion;
     private ApplicationManager mApplicationManager;
@@ -25,6 +25,7 @@ class OrbSession implements IOrbSession {
     private Bridge mBridge;
     private BrowserView mBrowserView;
     private DsmccClient mDsmccClient;
+    private String MOCK_NEW_MEDIA_ID = "urn:broadcaster:programme:1249863457643";
 
     /**
      * TV browser session.
@@ -313,6 +314,7 @@ class OrbSession implements IOrbSession {
             }
         });
     }
+
 
     /**
      * Get the View of the TV browser session. This should be added to the content view of the
@@ -1207,15 +1209,237 @@ class OrbSession implements IOrbSession {
 
     /**
      * @since 204
-     * 
+     *
      * Request for the Description of the current media playback on the application
      */
     @Override
-    public void onRequestMediaDescription() {
+    public boolean onVoiceRequestDescription() {
         if (mOrbHbbTVVersion < 204) {
             throw new UnsupportedOperationException("Unsupported 204 API.");
         }
 
-        mJsonRpc.onRequestMediaDescription();
+        mOrbSessionCallback.consoleLog("Request for the information of media playing...");
+        mJsonRpc.onVoiceRequestDescription();
+        return true;
+    }
+
+    /**
+     * @since 204
+     *
+     * Request to deliver a text input, from voice command, to applications
+     *
+     * @param input The content of the text
+     * @return true if this event has been handled, and false if not
+     */
+    @Override
+    public boolean onVoiceRequestTextInput(String input) {
+        if (mOrbHbbTVVersion < 204) {
+            throw new UnsupportedOperationException("Unsupported 204 API.");
+        }
+
+        // Mock function to display the input text
+        dispatchTextInput(input);
+        mOrbSessionCallback.consoleLog("Enter text {" + input + "}");
+        return true;
+    }
+
+    /**
+     * @since 204
+     *
+     * Called to send an intent, from a voice command, to applications
+     *
+     * @param action The index number of the intent, from intent.media.pause to intent.playback
+     * @param info   The value uniquely identifying a piece of content:
+     *               - INTENT_MEDIA_SEEK_WALLCLOCK: a wall clock time
+     *               - INTENT_DISPLAY: a URI
+     *               - INTENT_SEARCH: a search term specified by the user.
+     *               - INTENT_PLAYBACK: a URI
+     * @param anchor The value indicates an anchor point of the content, which is either "start" or "end"
+     * @param offset The number value for the time position, a number of seconds
+     * @return true if this event has been handled, and false if not
+     */
+    @Override
+    public boolean onVoiceSendIntent(Integer action, String info, String anchor,
+                                     int offset) {
+        if (mOrbHbbTVVersion < 204) {
+            throw new UnsupportedOperationException("Unsupported 204 API.");
+        }
+
+        String mediaId;
+        switch (action) {
+            case INTENT_MEDIA_PAUSE:
+                onSendIntentMediaBasics(INTENT_MEDIA_PAUSE);
+                mOrbSessionCallback.consoleLog("Send an intent, action: pause");
+                return true;
+            case INTENT_MEDIA_PLAY:
+                onSendIntentMediaBasics(INTENT_MEDIA_PLAY);
+                mOrbSessionCallback.consoleLog("Send an intent, action: play");
+                return true;
+            case INTENT_MEDIA_FAST_FORWARD:
+                onSendIntentMediaBasics(INTENT_MEDIA_FAST_FORWARD);
+                mOrbSessionCallback.consoleLog("Send an intent, action: fast-forward");
+                return true;
+            case INTENT_MEDIA_FAST_REVERSE:
+                onSendIntentMediaBasics(INTENT_MEDIA_FAST_REVERSE);
+                mOrbSessionCallback.consoleLog("Send an intent, action: fast-reverse");
+                return true;
+            case INTENT_MEDIA_STOP:
+                onSendIntentMediaBasics(INTENT_MEDIA_STOP);
+                mOrbSessionCallback.consoleLog("Send an intent, action: stop");
+                return true;
+            case INTENT_MEDIA_SEEK_CONTENT:
+                if (anchor.equals("start") || anchor.equals("end")) {
+                    onSendIntentMediaSeekContent(anchor, offset);
+                    mOrbSessionCallback.consoleLog("Send an intent, action: seek-content");
+                    return true;
+                }
+                break;
+            case INTENT_MEDIA_SEEK_RELATIVE:
+                onSendIntentMediaSeekRelative(offset);
+                mOrbSessionCallback.consoleLog("Send an intent, action: seek-relative");
+                return true;
+            case INTENT_MEDIA_SEEK_LIVE:
+                onSendIntentMediaSeekLive(offset);
+                mOrbSessionCallback.consoleLog("Send an intent, action: seek-live");
+                return true;
+            case INTENT_DISPLAY:
+                // TODO - get mediaId by media name
+                mediaId = MOCK_NEW_MEDIA_ID;
+                onSendIntentDisplay(mediaId);
+                mOrbSessionCallback.consoleLog("Send an intent, action: display");
+                return true;
+            case INTENT_MEDIA_SEEK_WALLCLOCK:
+                onSendIntentMediaSeekWallclock(info);
+                mOrbSessionCallback.consoleLog("Send an intent, action: seek-wallclock");
+                return true;
+            case INTENT_SEARCH:
+                onSendIntentSearch(info);
+                mOrbSessionCallback.consoleLog("Send an intent, action: search");
+                return true;
+            case INTENT_PLAYBACK:
+                mediaId = MOCK_NEW_MEDIA_ID;
+                onSendIntentPlayback(mediaId, anchor, offset);
+                mOrbSessionCallback.consoleLog("Send an intent, action: playback");
+                return true;
+        }
+        return false;
+    }
+
+    /**
+     * @since 204
+     *
+     * Called to send a send a keyUp event, from a voice command, to the application, potentially dispatching the event and show a message on window log.
+     *
+     * @param action The index number of the intent, either pressing a button or showing a log
+     */
+    @Override
+    public boolean onVoiceSendKeyAction(Integer action) {
+        if (mOrbHbbTVVersion < 204) {
+            throw new UnsupportedOperationException("Unsupported 204 API.");
+        }
+
+        String buttonName = ACT_BUTTON_NAMES.getOrDefault(action, "invalid");
+        if (buttonName.equals("invalid")) {
+            return false;
+        }
+        mOrbSessionCallback.consoleLog("Press " + buttonName + " button");
+        int keyCode;
+        switch (action) {
+            case ACT_PRESS_BUTTON_NUMB_ZERO:
+            case ACT_PRESS_BUTTON_NUMB_ONE:
+            case ACT_PRESS_BUTTON_NUMB_TWO:
+            case ACT_PRESS_BUTTON_NUMB_THREE:
+            case ACT_PRESS_BUTTON_NUMB_FOUR:
+            case ACT_PRESS_BUTTON_NUMB_FIVE:
+            case ACT_PRESS_BUTTON_NUMB_SIX:
+            case ACT_PRESS_BUTTON_NUMB_SEVEN:
+            case ACT_PRESS_BUTTON_NUMB_EIGHT:
+            case ACT_PRESS_BUTTON_NUMB_NINE:
+                keyCode = KeyEvent.KEYCODE_0 + action - ACT_PRESS_BUTTON_NUMB_ZERO;
+                break;
+            case ACT_PRESS_BUTTON_RED:
+                keyCode = KeyEvent.KEYCODE_PROG_RED;
+                break;
+            case ACT_PRESS_BUTTON_GREEN:
+                keyCode = KeyEvent.KEYCODE_PROG_GREEN;
+                break;
+            case ACT_PRESS_BUTTON_YELLOW:
+                keyCode = KeyEvent.KEYCODE_PROG_YELLOW;
+                break;
+            case ACT_PRESS_BUTTON_BLUE:
+                keyCode = KeyEvent.KEYCODE_PROG_BLUE;
+                break;
+            case ACT_PRESS_BUTTON_UP:
+                keyCode = KeyEvent.KEYCODE_DPAD_UP;
+                break;
+            case ACT_PRESS_BUTTON_DOWN:
+                keyCode = KeyEvent.KEYCODE_DPAD_DOWN;
+                break;
+            case ACT_PRESS_BUTTON_LEFT:
+                keyCode = KeyEvent.KEYCODE_DPAD_LEFT;
+                break;
+            case ACT_PRESS_BUTTON_RIGHT:
+                keyCode = KeyEvent.KEYCODE_DPAD_RIGHT;
+                break;
+            case ACT_PRESS_BUTTON_ENTER:
+                keyCode = KeyEvent.KEYCODE_ENTER;
+                break;
+            case ACT_PRESS_BUTTON_BACK:
+                keyCode = KeyEvent.KEYCODE_DEL;
+                break;
+            default:
+                return false;
+        }
+        KeyEvent event = new KeyEvent(KeyEvent.ACTION_UP, keyCode);
+        return dispatchKeyEvent(event);
+    }
+
+    /**
+     * @since 204
+     *
+     * Sends voice commands based on provided actions, messages, anchors, and offsets, where some of the parameters are optional.
+     *
+     * @param action The predefined index number of the intent, from intent.media.pause to intent.playback
+     * @param info   The value uniquely identifying a piece of content:
+     *               - INTENT_MEDIA_SEEK_WALLCLOCK: a wall clock time
+     *               - INTENT_DISPLAY: a URI
+     *               - INTENT_SEARCH: a search term specified by the user.
+     *               - INTENT_PLAYBACK: a URI
+     * @param anchor The value indicates an anchor point of the content...
+     * @param offset The number value for the time position, a number of seconds
+     * @return True if the command is successfully executed; otherwise, handles appropriately.
+     */
+    public boolean sendVoiceCommand(Integer action, String info, String anchor, int offset) {
+        if (mOrbHbbTVVersion < 204) {
+            throw new UnsupportedOperationException("Unsupported 204 API.");
+        }
+
+        switch (action) {
+            case INTENT_MEDIA_PAUSE:
+            case INTENT_MEDIA_PLAY:
+            case INTENT_MEDIA_FAST_FORWARD:
+            case INTENT_MEDIA_FAST_REVERSE:
+            case INTENT_MEDIA_STOP:
+            case INTENT_MEDIA_SEEK_CONTENT:
+            case INTENT_MEDIA_SEEK_RELATIVE:
+            case INTENT_MEDIA_SEEK_LIVE:
+            case INTENT_MEDIA_SEEK_WALLCLOCK:
+            case INTENT_SEARCH:
+            case INTENT_DISPLAY:
+            case INTENT_PLAYBACK:
+                return onVoiceSendIntent(action, info, anchor, offset);
+            case ACT_REQUEST_MEDIA_DESCRIPTION:
+                return onVoiceRequestDescription();
+            case ACT_REQUEST_TEXT_INPUT:
+                return onVoiceRequestTextInput(info);
+            case LOG_MESSAGE:
+            case LOG_ERROR_NONE_ACTION:
+            case LOG_ERROR_MULTI_ACTIONS:
+            case LOG_ERROR_INTENT_SEND:
+                mOrbSessionCallback.consoleLog(info);
+                return true;
+            default:
+                return onVoiceSendKeyAction(action);
+        }
     }
 }
